refactor : AttendanceScheduler -> AttendanceEvent로 이름 변경
