+++ waldreg/service-layer/attendance/src/main/java/org/waldreg/attendance/event/publish/AttendanceStopEvent.java
@@ -1,4 +1,4 @@
-package org.waldreg.attendance.schedule.publish;
+package org.waldreg.attendance.event.publish;
 
 public final class AttendanceStopEvent{
