+++ waldreg/service-layer/attendance/src/main/java/org/waldreg/attendance/event/subscribe/AttendanceLeftTimeEvent.java
@@ -1,4 +1,4 @@
-package org.waldreg.attendance.schedule.subscribe;
+package org.waldreg.attendance.event.subscribe;
 
 public final class AttendanceLeftTimeEvent{
